Replace CleanupMiddleware with SentryContextMiddleware
Since we already register the ErrorHandler, the middleware shouldn't be
responsible for reporting errors anymore.

Its responsibilities are:
1. Clone the main hub to the current thread. This will make sure Sentry
   is accessible from inside a job so the users can add custom contexts.
2. Add some default context data to the scope, like the transaction name
   and queue name.
3. Clear the scope after the job is processed successfully.
