+++ sentry-sidekiq/lib/sentry-sidekiq.rb
@@ -2,7 +2,7 @@ require "sidekiq"
 require "sentry-ruby"
 require "sentry/sidekiq/version"
 require "sentry/sidekiq/error_handler"
-require "sentry/sidekiq/cleanup_middleware"
+require "sentry/sidekiq/sentry_context_middleware"
 # require "sentry/sidekiq/configuration"
 
 module Sentry
@@ -18,7 +18,7 @@ end
 Sidekiq.configure_server do |config|
   config.error_handlers << Sentry::Sidekiq::ErrorHandler.new
   config.server_middleware do |chain|
-    chain.add Sentry::Sidekiq::CleanupMiddleware
+    chain.add Sentry::Sidekiq::SentryContextMiddleware
   end
 end
